requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
pypdfium2==4.30.0
easyocr==1.7.2
numpy==2.1.2
pandas==2.2.3
python-multipart==0.0.9
rich==13.9.2
loguru==0.7.2
Pillow==11.0.0

